// nestjs-template/template/src/common/enums/config.enum.ts
export enum ConfigEnum {
  APP_PORT = 'APP_PORT',
  APP_PREFIX = 'APP_PREFIX',
  APP_VERSION = 'APP_VERSION',
  LOG_ON = 'LOG_ON',
  DATABASE_TYPE = 'DATABASE_TYPE',
  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_PORT = 'DATABASE_PORT',
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_PASSWORD = 'DATABASE_PASSWORD',
  DATABASE_NAME = 'DATABASE_NAME',
  DATABASE_SYNC = 'DATABASE_SYNC',
  DATABASE_LOGGING = 'DATABASE_LOGGING',
  DATABASE_SCHEMA = 'DATABASE_SCHEMA',
  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRES = 'JWT_EXPIRES',
  JWT_EXPIRES_REFRESH = 'JWT_EXPIRES_REFRESH',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_DB = 'REDIS_DB',
  REDIS_PASSWORD = 'REDIS_PASSWORD',
  REDIS_TYPE = 'REDIS_TYPE',
  REDIS_CLUSTER_HOST = 'REDIS_CLUSTER_HOST',
  REDIS_CLUSTER_PORT = 'REDIS_CLUSTER_PORT',
  CACHE_TYPE = 'CACHE_TYPE',
  CACHE_TTL = 'CACHE_TTL',
  CACHE_MAX_ITEMS = 'CACHE_MAX_ITEMS',
  MAIL_ON = 'MAIL_ON',
  QUEUE_ON = 'QUEUE_ON',
  CRON_ON = 'CRON_ON',
  APP_I18N = 'APP_I18N',
  APP_LANGUAGE = 'APP_LANGUAGE',
  APP_LANGUAGE_LIST = 'APP_LANGUAGE_LIST',
  APP_LANGUAGE_DEFAULT = 'APP_LANGUAGE_DEFAULT',
  SSH_HOST = 'SSH_HOST',
  SSH_USERNAME = 'SSH_USERNAME',
  SSH_PASSWORD = 'SSH_PASSWORD',
  SSH_PORT = 'SSH_PORT',
  SMTP_HOST = 'SMTP_HOST',
  SMTP_PORT = 'SMTP_PORT',
  SMTP_USER = 'SMTP_USER',
  SMTP_PASS = 'SMTP_PASS',
  ROLE_WHITELIST = 'ROLE_WHITELIST',
  WX_APPID = 'WX_APPID',
  WX_SECRET = 'WX_SECRET'
}